Fix for manual mode of OverlayTrigger.
Before this fix it was not possible to refer to the component wrapped by the OverlayTrigger using the ref property, because it was not cloned using cloneWithProps.

Now, in manuel mode,  we just return the unmodified child which will still have the ref and all children.
(cherry picked from commit 4e54431)