Rename "navOpen" to "navExpanded"
There seems to be a non-working mix navOpen and navExpanded which means
getInitialState() sets navExpanded in the state, but isNavOpen refers to
the uninitialised navOpen. I have made it all use navExpanded, which
appears to be the intention, and also added some sensible defaults to
make it easier to get right.