Adds Property to Modal for Large Size
Pretty simple alteration I found myself needing and couldn't find the best route to insert.

It looks like div.modal-dialog classes are not accessible from the Modal component.

In order to create large modals in Bootstrap, the class .modal-lg needs to be added to the div.modal-dialog element.

This adds a property to Modal "large" (Boolean) that will add that class to dialogClasses.