Enable loose option for Babel
enable loose, mode for Babel. This transpiles with a slightly lower
level of spec "correctness" but in the vast majority of cases this
doesn't matter, and leads to better performance. Admittedly I am not
backing that up with numbers, but there is almost no downside to
enabling it. Specifically though it doesn't use DefineProperty in
classes which is notoriously slow

http://babeljs.io/docs/advanced/loose/