decouple the Popover components from the base component
start work on decoupling and modulating the popover components
to their own respective components. This will allow the users
to choose how they want the popovers to be displayed, whether they
want a title, no title, or if they want to inject their own
custom title.

also start work on changing the popover docs to
reflect this new behavior.

also start work on updating the popover unit tests to reflect this
new behavior