Refactor ThemeProvider via hooks
The ThemeProvider should be refactored to a functional component
with hooks. I decided to useMemo() the `prefixes` value, to avoid
unnecessary computations if the value doesn't change between
renders.

`prefixes` was re-written to be an object instead of a Map object,
as it allows for a more simple shallow copy implementation.

Also, duplicate code was removed in createBootstrapComponent().