feat: add default variant to Alert/Badge
This sets a default variant and bg of "primary" for Alert and Badge
respectively, allowing them to be used without having to set a variant
or bg.

Bootstrap's docs on Alerts require the use of a contextual class
(https://getbootstrap.com/docs/5.1/components/alerts/) in order to have
proper styling.

This aligns Alert and Badge variant functionality with that of Button
and Navbar.

This could be viewed as a breaking change (e.g. for someone using Alert
without a variant [as it now adds one] or someone relying on Button's
`.btn-null` [if the variant was null]).  However, I'm inclined to view
this as a feature as it is adding & declaring functionality in the
public API where behaviour wasn't previously explicitly defined in
examples or docs.