docs: fixed mistake with grid example
Error in the ["Grid"](https://react-bootstrap.netlify.app/docs/layout/grid/#responsive-grids)

The example in the documentation suggest that the `order` property takes a value between `1` and `12` (inclusive). It is not working properly in the example.

The Bootstrap ["documentation"](https://getbootstrap.com/docs/5.3/utilities/flex/#order) says the range goes from `0` to `5` (inclusive).